requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
